Chore: Disallow as_conversions (#561)
Fix feature gated import in hash precompile


Change missed featured `into` to `try_into`


birchmd fixes


Cleaner usize conversion error


Fix missing vec in hash


Fix tests


Fix V check


Fix testnet env EVM account


Change hex feature from alloc to std


Improve u64 to u8 conversion in JSON


Cargo fmt

+++ engine-types/src/types/gas.rs
@@ -67,11 +67,11 @@ impl Add<EthGas> for EthGas {
     }
 }
 
-impl Div<usize> for EthGas {
+impl Div<u64> for EthGas {
     type Output = EthGas;
 
-    fn div(self, rhs: usize) -> Self::Output {
-        EthGas(self.0 / rhs as u64)
+    fn div(self, rhs: u64) -> Self::Output {
+        EthGas(self.0 / rhs)
     }
 }
 
@@ -79,7 +79,7 @@ impl Mul<EthGas> for u32 {
     type Output = EthGas;
 
     fn mul(self, rhs: EthGas) -> Self::Output {
-        EthGas(self as u64 * rhs.0)
+        EthGas(u64::from(self) * rhs.0)
     }
 }
 
@@ -87,15 +87,15 @@ impl Mul<u32> for EthGas {
     type Output = EthGas;
 
     fn mul(self, rhs: u32) -> Self::Output {
-        EthGas(self.0 * rhs as u64)
+        EthGas(self.0 * u64::from(rhs))
     }
 }
 
-impl Mul<usize> for EthGas {
+impl Mul<u64> for EthGas {
     type Output = EthGas;
 
-    fn mul(self, rhs: usize) -> Self::Output {
-        EthGas(self.0 * rhs as u64)
+    fn mul(self, rhs: u64) -> Self::Output {
+        EthGas(self.0 * rhs)
     }
 }
